backend/requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.2
dj-rest-auth==7.0.1
Django==4.2.21
django-allauth==65.8.1
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
ecdsa==0.19.1
google-auth==2.40.2
idna==3.10
oauthlib==3.2.2
psycopg2-binary==2.9.10
pyasn1==0.4.8
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
python-jose==3.4.0
requests==2.32.3
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
